More changes based on mt review
