+++ neqo-transport/src/qlog.rs
@@ -16,7 +16,7 @@ use crate::frame::{self, Frame};
 use crate::packet::{DecryptedPacket, PacketNumber, PacketType};
 use crate::path::Path;
 use crate::tparams::{self, TransportParametersHandler};
-use crate::{Res, DEFAULT_QUIC_VERSION};
+use crate::{QuicVersion, Res};
 
 pub fn connection_tparams_set(
     qlog: &mut Option<NeqoQlog>,
@@ -163,7 +163,7 @@ fn connection_started(qlog: &mut Option<NeqoQlog>, path: &Path) -> Res<()> {
             Some("QUIC".into()),
             path.local_address().port().into(),
             path.remote_address().port().into(),
-            Some(format!("{:x}", DEFAULT_QUIC_VERSION.as_u32())),
+            Some(format!("{:x}", QuicVersion::default().as_u32())),
             Some(format!("{}", path.local_cid())),
             Some(format!("{}", path.remote_cid())),
         ))?;
